Вынес endPoints в раздельные файлы.
