+++ src/boss_handlers.rs
@@ -1,40 +1,11 @@
 use iron::*;
 use iron::mime::{Mime, TopLevel, SubLevel};
-use fair_boss_drop_server::postgres::Connection;
 
 use std::io::Read;
 use std::sync::Mutex;
-use std::collections::LinkedList;
 
-use BdLayer::PostgresCommands::*;
-use BdLayer::ItemsCommands::*;
 use BdLayer::BossCommands::*;
 use BdLayer::PostgresDealer::*;
-
-
-pub fn get_item_types(sdb: &Mutex<PostgresSqlData>, req: &mut Request) -> IronResult<Response> {
-    let mut bd_data = sdb.lock().unwrap();
-
-    let mut getItemTypes = PostgresGetItemTypes::new();
-    match bd_data.doCommand(&mut getItemTypes) {
-        Ok(res) => {  println!("get item types"); },
-        Err(er) => {  println!("{}",er); }
-    }
-    println!("start {:?}",getItemTypes.getData());
-
-    Ok(Response::with((status::Ok,"Get Item Types executed")))
-}
-
-pub fn insert_item_type(sdb: &Mutex<PostgresSqlData>, req: &mut Request) -> IronResult<Response> {
-    let mut bd_data = sdb.lock().unwrap();
-
-    let mut addItemType = PostgresInsertItemType::new("tip1");
-    match bd_data.doCommand(&mut addItemType) {
-        Ok(res) => {  println!("item added"); },
-        Err(er) => {  println!("{}",er); }
-    };
-    Ok(Response::with((status::Ok,"insert item type executed")))
-}
 
 pub fn insert_boss(sdb: &Mutex<PostgresSqlData>, req: &mut Request) -> IronResult<Response> {
     let mut bd_data = sdb.lock().unwrap();
@@ -43,8 +14,8 @@ pub fn insert_boss(sdb: &Mutex<PostgresSqlData>, req: &mut Request) -> IronResul
     items.push(ItemProbability{ _id:2, _probability: 0.5});
     items.push(ItemProbability{ _id:3, _probability: 0.25});
 
-    let mut insertBoss = PostgresInsertBoss::new("boss4",2,items);
-    match bd_data.doCommand(&mut insertBoss) {
+    let mut insert_boss = PostgresInsertBoss::new("boss4",2,items);
+    match bd_data.doCommand(&mut insert_boss) {
         Ok(res) => {  println!("boss added"); },
         Err(er) => {  println!("{}",er); }
     }
@@ -55,8 +26,8 @@ pub fn insert_boss(sdb: &Mutex<PostgresSqlData>, req: &mut Request) -> IronResul
 pub fn get_boss(sdb: &Mutex<PostgresSqlData>, req: &mut Request) -> IronResult<Response> {
     let mut bd_data = sdb.lock().unwrap();
 
-    let mut getBoss = PostgresGetBoss::new("boss3");
-    match bd_data.doCommand(&mut getBoss) {
+    let mut get_boss = PostgresGetBoss::new("boss3");
+    match bd_data.doCommand(&mut get_boss) {
         Ok(res) => {  println!("get boss"); },
         Err(er) => {  println!("{}",er); }
     }
@@ -67,10 +38,10 @@ pub fn get_boss(sdb: &Mutex<PostgresSqlData>, req: &mut Request) -> IronResult<R
 pub fn get_bosses(sdb: &Mutex<PostgresSqlData>, req: &mut Request) -> IronResult<Response> {
     let mut bd_data = sdb.lock().unwrap();
 
-    let mut getBosses = PostgresGetBosses::new();
-    match bd_data.doCommand(&mut getBosses) {
+    let mut get_bosses = PostgresGetBosses::new();
+    match bd_data.doCommand(&mut get_bosses) {
         Ok(res) => {  println!("get bosses");
-                      println!("Bosses {:?}",getBosses.getData());
+                      println!("Bosses {:?}",get_bosses.getData());
         },
         Err(er) => {  println!("{}",er); }
     }
