+++ src/main.rs
@@ -10,7 +10,8 @@ extern crate iron;
 extern crate router;
 extern crate url;
 
-mod handlers;
+mod item_handlers;
+mod boss_handlers;
 
 use iron::*;
 use std::sync::{Arc, Mutex};
@@ -18,7 +19,6 @@ use std::sync::{Arc, Mutex};
 use fair_boss_drop_server::BdLayer;
 use BdLayer::PostgresDealer::*;
 use BdLayer::PostgresCommands::PostgresInitTables;
-use BdLayer::BossCommands::*;
 
 fn serve(db: PostgresSqlData) {
     let sdb = Arc::new(Mutex::new(db));
@@ -27,13 +27,13 @@ fn serve(db: PostgresSqlData) {
         let sdb_ = sdb.clone();
 
         router.get("/api/v0/itemtypes", move |req: &mut Request|
-                   handlers::get_item_types(&sdb_.clone(), req), "get_all_item_types");
+                   item_handlers::get_item_types(&sdb_.clone(), req), "get_all_item_types");
     }
     {
         let sdb_ = sdb.clone();
 
         router.get("/api/v0/bosses", move |req: &mut Request|
-                   handlers::get_bosses(&sdb_.clone(), req), "get_all_bosses");
+                   boss_handlers::get_bosses(&sdb_.clone(), req), "get_all_bosses");
     }
 
     Iron::new(router).http("localhost:3000").unwrap();
@@ -44,8 +44,8 @@ fn main() {
     let mut bd_data = PostgresSqlData::new();
     bd_data.connect().unwrap();
     {
-        let mut initTables = PostgresInitTables::new();
-        bd_data.doCommand(&mut initTables).expect("Error when init db tables");
+        let mut init_tables = PostgresInitTables::new();
+        bd_data.doCommand(&mut init_tables).expect("Error when init db tables");
     }
 
     serve(bd_data);
